+++ GCP/snippets/2026-07-14-list-gce-instances-and-gcs-buckets.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+# last_verified: 2026-07-14 - gcloud CLI n/a
+# I listed Compute Engine instances and Cloud Storage buckets
+
+gcloud compute instances list --format="table(name, zone, status, machineType)"
+gsutil ls